warnings on our own sources, and the one conversion they found
-Wconversion was predicted to be the high-yield flag here. Measured across
all seventeen of our sources: 113 warnings, 112 inside HAL/CMSIS headers
they merely include, one ours - board_selftest.c returning a `0U` ternary
as uint8_t. The integer-heavy files are heavy in explicit casts.

* -Wextra -Wshadow on all seventeen: zero hits, so free.
* -Wconversion on the ten that include no HAL header - Modbus/ entire,
  plus wire, cmd, cmd_board, cmd_test, link, link_report. Suppressing the
  HAL's share needs its includes marked SYSTEM, which lives in the CubeMX
  file and would not survive a regeneration.
* The subset is invariant 1's list plus the layers that stayed
  hardware-free, so the flag guards it: a HAL include there lights up.
* CubeMX's own sources are left alone.

Not -Werror: HAL noise would make it a build failure the moment anyone
touches a file that includes HAL.

Verified in compile_commands.json that the flags reach exactly those files.
Clean rebuild 0 warnings, flashed. Gate: 1257 passed, 0 failed, 7 of 7.

+++ Board/Src/board_selftest.c
@@ -42,7 +42,7 @@ static void add(board_check_t *out, uint8_t *n, uint8_t capacity,
 
 static uint8_t verdict(bool ok)
 {
-  return ok ? BOARD_CHECK_PASS : BOARD_CHECK_FAIL;
+  return (uint8_t)(ok ? BOARD_CHECK_PASS : BOARD_CHECK_FAIL);
 }
 
 /* Every configured channel should leave at most its own bit and, for a
